chore: move other file streaming code to stream file
